requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.1
python-dotenv==1.0.1
email-validator==2.2.0
openai